requirements.txt
regex==2026.1.15
certifi==2026.1.4
jsonschema-specifications==2025.9.1
tzdata==2025.3
pytz==2025.2
fsspec==2023.10.0
setuptools==65.5.0
pyzmq==27.1.0
packaging==26.0
attrs==25.4.0
aiofiles==24.1.0
pip==23.0.1
pyarrow==23.0.1
rich==14.3.2
pillow==12.1.1
ipython==8.12.3
jupyter-client==8.8.0
click==8.3.1
nbconvert==7.17.0
psutil==7.2.2
ipykernel==7.2.0
multidict==6.7.1
gradio==6.6.0
tornado==6.5.4
bleach==6.3.0
pyyaml==6.0.3
traitlets==5.14.3
nbformat==5.10.4
jupyter-core==5.9.1
decorator==5.2.1
async-timeout==5.0.1
tqdm==4.67.3
fonttools==4.61.1
transformers==4.36.2
jsonschema==4.26.0
typing-extensions==4.15.0
beautifulsoup4==4.14.3
anyio==4.12.1
platformdirs==4.9.2
pexpect==4.9.0
markdown-it-py==4.0.0
filelock==3.24.3
aiohttp==3.13.3
orjson==3.11.7
idna==3.11
matplotlib==3.10.8
xxhash==3.6.0
charset-normalizer==3.4.4
networkx==3.4.2
pyparsing==3.3.2
mistune==3.2.0
jinja2==3.1.6
prompt-toolkit==3.0.52
markupsafe==3.0.3
asttokens==3.0.1
pydantic-core==2.41.5
requests==2.32.5
fastjsonschema==2.21.2
pygments==2.19.2
datasets==2.16.1
pydantic==2.12.5
semantic-version==2.10.0
python-dateutil==2.9.0.post0
soupsieve==2.8.3
urllib3==2.6.3
aiohappyeyeballs==2.6.1
pandas==2.3.3
executing==2.2.1
torchaudio==2.1.2
torch==2.1.2
gradio-client==2.1.0
numpy==1.26.4
yarl==1.22.0
six==1.17.0
sympy==1.14.0
faiss-cpu==1.13.2
debugpy==1.8.20
frozenlist==1.8.0
nest-asyncio==1.6.0
shellingham==1.5.4
pandocfilters==1.5.1
kiwisolver==1.4.9
tinycss2==1.4.0
aiosignal==1.4.0
contourpy==1.3.2
exceptiongroup==1.3.1
mpmath==1.3.0
hf-xet==1.2.0
brotli==1.2.0
httpcore==1.0.9
ffmpy==1.0.0
fastapi==0.129.0
multiprocess==0.70.15
starlette==0.52.1
uvicorn==0.41.0
referencing==0.37.0
huggingface-hub==0.36.2
rpds-py==0.30.0
httpx==0.28.1
pydub==0.25.1
typer==0.24.0
jedi==0.19.2
torchvision==0.16.2
h11==0.16.0
tokenizers==0.15.2
tomlkit==0.13.3
cycler==0.12.1
nbclient==0.10.4
parso==0.8.6
pickleshare==0.7.5
defusedxml==0.7.1
ptyprocess==0.7.0
pyarrow-hotfix==0.7
safetensors==0.7.0
annotated-types==0.7.0
stack-data==0.6.3
docopt==0.6.2
wcwidth==0.6.0
webencodings==0.5.1
pipreqs==0.5.0
typing-inspection==0.4.2
propcache==0.4.1
dill==0.3.7
jupyterlab-pygments==0.3.0
pure-eval==0.2.3
comm==0.2.3
matplotlib-inline==0.2.1
backcall==0.2.0
yarg==0.1.9
safehttpx==0.1.7
appnope==0.1.4
groovy==0.1.2
mdurl==0.1.2
python-multipart==0.0.22
annotated-doc==0.0.4